Clear every remaining analyser finding that is safe to clear
Everything mechanical is done: unnecessary dict()/set()/list() calls,
literal concatenation rewritten as unpacking, redundant int() casts,
generator-to-comprehension, collapsible ifs, yoda conditions, f-strings
without placeholders, unsorted __all__, nested with statements,
suppressible exceptions, and an ambiguous multiplication sign in a
comment. Unpacked values nobody reads are now named _ so the next reader
knows that is deliberate, and a mutable default in a test fixture became a
module-level singleton.

Two loops no longer reassign the variable they iterate. The adoption loop
now returns early for anything that is not an adoption, so the enriched
copy has its own name instead of shadowing the planner's original.

Three categories are left, each for a reason that was checked rather than
assumed:

TRY004 (18) wants TypeError where these raise ValueError. The validation
layer raises ValueError uniformly for "this input is not acceptable", and
96 assertRaises(ValueError) in the tests plus seven except ValueError
handlers in the modules depend on that. Changing it would break the
contract to satisfy a preference.

PLW0603 (13) flags the global statement. Module-level run state is how
this program is built; removing it is an architecture rewrite, not a fix.

PLC0415 (49) wants the function-local imports hoisted. I tried it, and it
would have broken every run: one of them is IPython, which is absent
outside a notebook, is not in requirements.txt, and sits inside a try for
exactly that reason. Hoisting it to module scope makes the import fail in
CI. An earlier grep of mine had also suggested these were redundant - ^
import re matches "import requests" - and an AST check showed none of them
is imported at the top at all, so every one is load-bearing.

+++ screener_broker_sync.py
@@ -59,7 +59,7 @@ def _whole(value):
     number = _finite(value)
     if number is None:
         return None
-    rounded = int(round(number))
+    rounded = round(number)
     return rounded if abs(number - rounded) < 1e-6 else None
